refactor(charts): share the chart config the two charts have in common
Pulls what the dashboard's invoice chart and the project budget chart both
configure out of `invoicesChart` into `chartBase`, and the mounting and the
bold-hovered-label behaviour out of `UiChart` into `UiHighchart`, so the
project chart can reuse both.

The chart is redrawn from the function it is handed, so that function now
reads its series out where the computed can see it — a closure over the query
data alone never invalidates, and the chart kept the empty series it was first
mounted with.
